refactor: improve modularity, abstraction, and service layer separation
