+++ app/db/constants.py
@@ -1,2 +1,13 @@
 # Generic fields
 ID = "_id"
+
+CLASS_WINDOW_DAYS = 14
+CLASS_COLLECTION = "classes"
+
+class_name = "class_name"
+start_time = "start_time"
+end_time = "end_time"
+location = "location"
+capacity = "capacity"
+trainer_name = "trainer_name"
+remaining_spots = "remaining_spots"
